### 2018-09-14 aea5f96
Small fixes to JOSS paper
