### 2018-09-20 b320503
Replacing hard coded SPARQL language tagged queries with LanguageFilteringRDFService
Change-Id: I0c557658cefc9537411082680e53ef378f4cddca
