+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/visualization/utilities/VisualizationCaches.java
@@ -99,7 +99,7 @@ final public class VisualizationCaches {
             new CachingRDFServiceExecutor<>(
                     new CachingRDFServiceExecutor.RDFServiceCallable<Map<String, Person>>(visualizationAffinity) {
                         @Override
-                        protected Map<String, Person> callWithService(RDFService rdfService, Locale locale) throws Exception {
+                        protected Map<String, Person> callWithService(RDFService rdfService) throws Exception {
                             final Map<String, Person> map = new HashMap<String, Person>();
 
                             String construct = QueryConstants.getSparqlPrefixQuery() +
@@ -232,14 +232,13 @@ final public class VisualizationCaches {
             new CachingRDFServiceExecutor<>(
                     new CachingRDFServiceExecutor.RDFServiceCallable<Map<String, String>>(visualizationAffinity) {
                         @Override
-                        protected Map<String, String> callWithService(RDFService rdfService, Locale locale) throws Exception {
+                        protected Map<String, String> callWithService(RDFService rdfService) throws Exception {
                             String query = QueryConstants.getSparqlPrefixQuery() +
                                     "SELECT ?org ?orgLabel\n" +
                                     "WHERE\n" +
                                     "{\n" +
                                     "  ?org a foaf:Organization .\n" +
                                     "  ?org rdfs:label ?orgLabel .\n" +
-                                    "    FILTER(langMatches( lang(?orgLabel), '"+locale.toLanguageTag()+"' ))"+
                                     "}\n";
 
                             final Map<String, String> map = new HashMap<>();
@@ -266,7 +265,7 @@ final public class VisualizationCaches {
             new CachingRDFServiceExecutor<>(
                     new CachingRDFServiceExecutor.RDFServiceCallable<Map<String, Set<String>>>(visualizationAffinity) {
                         @Override
-                        protected Map<String, Set<String>> callWithService(RDFService rdfService, Locale locale) throws Exception {
+                        protected Map<String, Set<String>> callWithService(RDFService rdfService) throws Exception {
                             String query = QueryConstants.getSparqlPrefixQuery() +
                                     "SELECT ?org ?subOrg\n" +
                                     "WHERE\n" +
@@ -306,7 +305,7 @@ final public class VisualizationCaches {
             new CachingRDFServiceExecutor<>(
                     new CachingRDFServiceExecutor.RDFServiceCallable<Map<String, String>>() {
                         @Override
-                        protected Map<String, String> callWithService(RDFService rdfService, Locale locale) throws Exception {
+                        protected Map<String, String> callWithService(RDFService rdfService) throws Exception {
                             String query = QueryConstants.getSparqlPrefixQuery() +
                                     "SELECT ?org ?typeLabel\n" +
                                     "WHERE\n" +
@@ -314,7 +313,6 @@ final public class VisualizationCaches {
                                     "    ?org a foaf:Organization .\n" +
                                     "    ?org vitro:mostSpecificType ?type .\n" +
                                     "    ?type rdfs:label ?typeLabel .\n" +
-                                    "    FILTER(langMatches( lang(?typeLabel), '"+locale.toLanguageTag()+"' ))"+
                                     "}\n";
 
                             final Map<String, String> map = new HashMap<>();
@@ -340,7 +338,7 @@ final public class VisualizationCaches {
             new CachingRDFServiceExecutor<OrganizationPeopleMap>(
                     new CachingRDFServiceExecutor.RDFServiceCallable<OrganizationPeopleMap>(visualizationAffinity) {
                         @Override
-                        protected OrganizationPeopleMap callWithService(RDFService rdfService, Locale locale) throws Exception {
+                        protected OrganizationPeopleMap callWithService(RDFService rdfService) throws Exception {
[…]
