+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/visualization/temporalgraph/TemporalGrantVisualizationRequestHandler.java
@@ -7,7 +7,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 
@@ -43,12 +42,12 @@ import edu.cornell.mannlib.vitro.webapp.visualization.visutils.VisualizationRequ
 
 public class TemporalGrantVisualizationRequestHandler implements
 		VisualizationRequestHandler {
-	
+
 	@Override
 	public ResponseValues generateStandardVisualization(
 			VitroRequest vitroRequest, Log log, Dataset dataset)
 			throws MalformedQueryParametersException {
-		
+
 		String entityURI = vitroRequest.getParameter(VisualizationFrameworkConstants.INDIVIDUAL_URI_KEY);
 		return generateStandardVisualizationForGrantTemporalVis(vitroRequest, log, dataset, entityURI);
 	}
@@ -61,15 +60,15 @@ public class TemporalGrantVisualizationRequestHandler implements
 		if (StringUtils.isBlank(entityURI)) {
 			entityURI = OrganizationUtilityFunctions.getStaffProvidedOrComputedHighestLevelOrganization(
 										log,
-										dataset, 
+										dataset,
 										vitroRequest);
-			
+
 		}
-		
+
 		return prepareStandaloneMarkupResponse(vitroRequest, entityURI);
 	}
-	
-	
+
+
 	@Override
 	public ResponseValues generateVisualizationForShortURLRequests(
 			Map<String, String> parameters, VitroRequest vitroRequest, Log log,
@@ -77,7 +76,7 @@ public class TemporalGrantVisualizationRequestHandler implements
 
 		return generateStandardVisualizationForGrantTemporalVis(
 				vitroRequest, log, dataset, parameters.get(VisualizationFrameworkConstants.INDIVIDUAL_URI_KEY));
-		
+
 	}
 
 	@Override
@@ -85,29 +84,29 @@ public class TemporalGrantVisualizationRequestHandler implements
 			throws MalformedQueryParametersException, JsonProcessingException {
 
 		String entityURI = vitroRequest.getParameter(VisualizationFrameworkConstants.INDIVIDUAL_URI_KEY);
-		
+
 		VisConstants.DataVisMode currentDataMode = VisConstants.DataVisMode.CSV;
-		
+
 		/*
-		 * This will provide the data in json format mainly used for standalone temporal vis. 
+		 * This will provide the data in json format mainly used for standalone temporal vis.
 		 * */
 		if (VisualizationFrameworkConstants.JSON_OUTPUT_FORMAT
 					.equalsIgnoreCase(vitroRequest
 							.getParameter(VisualizationFrameworkConstants.VIS_MODE_KEY))) {
-			
+
 			currentDataMode = VisConstants.DataVisMode.JSON;
-			
+
 			if (StringUtils.isBlank(entityURI)) {
-				
+
 				entityURI = OrganizationUtilityFunctions
 								.getStaffProvidedOrComputedHighestLevelOrganization(
 										log,
 										dataset,
-										vitroRequest);								
-								
-			} 
-			
-		} 		
+										vitroRequest);
+
+			}
+
+		}
 
 		try {
 			return getSubjectEntityAndGenerateDataResponse(
@@ -120,25 +119,25 @@ public class TemporalGrantVisualizationRequestHandler implements
 			VisualizationCaches.buildMissing();
 		}
 	}
-	
+
 	private Map<String, String> prepareDataErrorResponse() {
 		String outputFileName = "no-organization_grants-per-year.csv";
-		
+
 		Map<String, String> fileData = new HashMap<String, String>();
-		
+
[…]

[…]
