### 2018-09-25 d757ec2
Implementation of Advanced Role Management, including migration of existing roles
