+++ api/src/main/java/edu/cornell/mannlib/vivo/auth/policy/PresentationChecker.java
@@ -7,6 +7,7 @@ import java.util.List;
 import edu.cornell.mannlib.vitro.webapp.auth.policy.ifaces.PolicyDecision;
 import edu.cornell.mannlib.vitro.webapp.auth.policy.specialrelationships.RelationshipChecker;
 import edu.cornell.mannlib.vitro.webapp.auth.requestedAction.propstmt.AbstractPropertyStatementAction;
+import org.apache.jena.ontology.OntModel;
 
 /**
  * Does the requested action involve a change to a Presentation that the
@@ -18,41 +19,30 @@ public class PresentationChecker extends RelationshipChecker {
 	private static final String URI_PRESENTER_ROLE_TYPE = NS_CORE
 			+ "PresenterRole";
 
-	private final String[] resourceUris;
-
-	public PresentationChecker(AbstractPropertyStatementAction action) {
-		super(action.getOntModel());
-		this.resourceUris = action.getResourceUris();
-	}
-
 	/**
 	 * A self-editor is authorized to add, edit, or delete a statement if the
 	 * subject or object refers to a Presentation, and if the self-editor:
 	 * 
 	 * 1) is a Presenter of that Presentation
 	 */
-	public PolicyDecision isAuthorized(List<String> userUris) {
-		for (String resourceUri : resourceUris) {
-			if (isPresentation(resourceUri)) {
-				if (anyUrisInCommon(userUris, getUrisOfPresenters(resourceUri))) {
-					return authorizedPresenter(resourceUri);
+	public boolean isRelated(OntModel ontModel, List<String> fromUris, List<String> toUris) {
+		for (String resourceUri : fromUris) {
+			if (isPresentation(ontModel, resourceUri)) {
+				if (anyUrisInCommon(ontModel, toUris, getUrisOfPresenters(ontModel, resourceUri))) {
+					return true;
 				}
 			}
 		}
-		return null;
+
+		return false;
 	}
 
-	private boolean isPresentation(String resourceUri) {
-		return isResourceOfType(resourceUri, URI_PRESENTATION_TYPE);
+	private boolean isPresentation(OntModel ontModel, String resourceUri) {
+		return isResourceOfType(ontModel, resourceUri, URI_PRESENTATION_TYPE);
 	}
 
-	private List<String> getUrisOfPresenters(String resourceUri) {
-		return getObjectsThroughLinkingNode(resourceUri, URI_REALIZES,
+	private List<String> getUrisOfPresenters(OntModel ontModel, String resourceUri) {
+		return getObjectsThroughLinkingNode(ontModel, resourceUri, URI_REALIZES,
 				URI_PRESENTER_ROLE_TYPE, URI_INHERES_IN);
 	}
-
-	private PolicyDecision authorizedPresenter(String resourceUri) {
-		return authorizedDecision("User is a Presenter of " + resourceUri);
-	}
-
 }

[…]
