### 2018-09-25 47c0147
Implementation of Advanced Role Management, including migration of existing roles
