### 2018-09-25 ac0814f
Implementation of Advanced Role Management, including migration of existing roles
