+++ api/src/main/java/edu/cornell/mannlib/vivo/auth/policy/InfoContentEntityChecker.java
@@ -7,6 +7,7 @@ import java.util.List;
 import edu.cornell.mannlib.vitro.webapp.auth.policy.ifaces.PolicyDecision;
 import edu.cornell.mannlib.vitro.webapp.auth.policy.specialrelationships.RelationshipChecker;
 import edu.cornell.mannlib.vitro.webapp.auth.requestedAction.propstmt.AbstractPropertyStatementAction;
+import org.apache.jena.ontology.OntModel;
 
 /**
  * Does the requested action involve a change to an Info Content Entity that the
@@ -17,13 +18,6 @@ public class InfoContentEntityChecker extends RelationshipChecker {
 	private static final String URI_FEATURES_PROPERTY = NS_CORE + "features";
 	private static final String URI_AUTHORSHIP_TYPE = NS_CORE + "Authorship";
 	private static final String URI_EDITORSHIP_TYPE = NS_CORE + "Editorship";
-
-	private final String[] resourceUris;
-
-	public InfoContentEntityChecker(AbstractPropertyStatementAction action) {
-		super(action.getOntModel());
-		this.resourceUris = action.getResourceUris();
-	}
 
 	/**
 	 * A self-editor is authorized to add, edit, or delete a statement if the
@@ -36,57 +30,45 @@ public class InfoContentEntityChecker extends RelationshipChecker {
 	 * 
 	 * 3) is Featured in that Info Content Entity.
 	 */
-	public PolicyDecision isAuthorized(List<String> userUris) {
-		for (String resourceUri : resourceUris) {
-			if (isInfoContentEntity(resourceUri)) {
-				if (anyUrisInCommon(userUris, getUrisOfAuthors(resourceUri))) {
-					return authorizedAuthor(resourceUri);
+	public boolean isRelated(OntModel ontModel, List<String> fromUris, List<String> toUris) {
+		for (String resourceUri : fromUris) {
+			if (isInfoContentEntity(ontModel, resourceUri)) {
+				if (anyUrisInCommon(ontModel, toUris, getUrisOfAuthors(ontModel, resourceUri))) {
+					return true;
 				}
-				if (anyUrisInCommon(userUris, getUrisOfEditors(resourceUri))) {
-					return authorizedEditor(resourceUri);
+				if (anyUrisInCommon(ontModel, toUris, getUrisOfEditors(ontModel, resourceUri))) {
+					return true;
 				}
-				if (anyUrisInCommon(userUris, getUrisOfFeatured(resourceUri))) {
-					return authorizedFeatured(resourceUri);
+				if (anyUrisInCommon(ontModel, toUris, getUrisOfFeatured(ontModel, resourceUri))) {
+					return true;
 				}
 			}
 		}
-		return null;
+
+		return false;
 	}
 
-	private boolean isInfoContentEntity(String resourceUri) {
-		return isResourceOfType(resourceUri, URI_INFO_CONTENT_TYPE);
+	private boolean isInfoContentEntity(OntModel ontModel, String resourceUri) {
+		return isResourceOfType(ontModel, resourceUri, URI_INFO_CONTENT_TYPE);
 	}
 
-	private List<String> getUrisOfEditors(String resourceUri) {
-		List<String> allRelatedUris = getObjectsThroughLinkingNode(resourceUri,
+	private List<String> getUrisOfEditors(OntModel ontModel, String resourceUri) {
+		List<String> allRelatedUris = getObjectsThroughLinkingNode(ontModel, resourceUri,
 				URI_RELATED_BY, URI_EDITORSHIP_TYPE, URI_RELATES);
 		// The editorship relates to the editors and to the resource itself.
 		allRelatedUris.remove(resourceUri);
 		return allRelatedUris;
 	}
 
[…]
