### 2018-09-26 f6eb0eb
Update default startup_listeners.txt for ElasticSearch
Related to: https://jira.duraspace.org/browse/VIVO-1600
