### 2018-10-02 #943 @jasonclark
@arfon Yes, I can edit this submission.
@mconlon17 Thanks for the submission. The software has a general description, a BSD 3-Clause "New" or "Revised" License, you are a recognized contributor on the project, and it is scoped correctly as research software (E.g., software for a research application). These are a few of my pre-review checks. Our next step is to find a reviewer.
