### 2018-10-04 #943 @jaynanavati
I am so sorry. I am busy with other pre-planned things. So, I am not
available.

On Wed, Oct 3, 2018 at 11:22 AM Jason A. Clark <notifications@github.com>
wrote:

> @gabrielspadon <https://github.com/gabrielspadon> Would you be available
> to review the VIVO software? We would be looking for a review in the next
> few weeks. Thanks for considering. cc @mconlon17
> <https://github.com/mconlon17>
>
