### 2018-10-26 167e33d
Improvements requested in PR review
