### 2018-10-26 aa59c57
Improvements requested in PR review
