### 2018-10-26 c822e77
Improvements requested in PR review
