### 2018-10-31 591e8c4
Add Travis-CI Badge to README.md
Resolves: https://jira.duraspace.org/browse/VIVO-1618
