### 2018-11-20 f5b6a27
Address code review comments
From:
- Vitro: https://github.com/vivo-project/Vitro/pull/80
- VIVO: https://github.com/vivo-project/VIVO/pull/85

Resolves:
https://github.com/vivo-project/Vitro/pull/80/files#r227428235
https://github.com/vivo-project/Vitro/pull/80/files#r227892906
https://github.com/vivo-project/Vitro/pull/80/files#r227905400
https://github.com/vivo-project/Vitro/pull/80/files#r227907884
https://github.com/vivo-project/Vitro/pull/80/files#r227909775
https://github.com/vivo-project/Vitro/pull/80/files#r227913953
https://github.com/vivo-project/Vitro/pull/80/files#r227913145
https
[…]
