### 2018-11-29 #943 @sonal-raj
> @sonal-raj We could really use a reviewer on this software. Thought your experience with Graphs and Graph Databases would be super useful here. Would you be available to review the VIVO software? We would be looking for a review in the next few weeks. Thanks for considering. cc @mconlon17 @arfon

@jasonclark Thanks for the request. I like the idea of this software and would love to help review it. Only constraint is my schedule is packed till 18th of Dec, but I am available post that. Depending on the urgency, if the time looks fine to you, I can definitely help review VIVO. Do let me know.
