### 2018-12-04 9a58cb1
Use diamond operator on ArrayList declarations
