### 2018-12-04 a0f5bf3
Use the diamond operator to simpplify declarations
