### 2018-12-05 0fa67e5
Updating AGROVOC REST API base URL.
* Resolves with https://jira.duraspace.org/browse/VIVO-1655
