### 2018-12-05 f2021ca
[VIVO-1656] Extension mechansim for passing properties to templates for Individuals
