+++ webapp/src/main/webapp/WEB-INF/resources/startup_listeners.txt
@@ -86,6 +86,9 @@ edu.cornell.mannlib.vitro.webapp.dao.jena.VClassGroupCache$Setup
 
 edu.cornell.mannlib.vivo.orcid.OrcidContextSetup
 
+# Register the extensions to the IndividualResponseBuilder (e.g. alt metric, plum analytics)
+edu.cornell.mannlib.vitro.webapp.controller.individual.VIVOIndividualResponseBuilderExtension$Setup
+
 # This should be near the end, because it will issue a warning if the connection to Solr times out.
 edu.cornell.mannlib.vitro.webapp.servlet.setup.SolrSmokeTest
 
