### 2018-12-05 fda5e83
[VIVO 1647] - Improve AgrovocService.java using diamond operator (#97)
* Use diamond operator on ArrayList declarations
* Use the diamond operator to simpplify declarations

Resolves: https://jira.duraspace.org/browse/VIVO-1647
