### 2018-12-05 94fcd23
[VIVO-1415] Claiming interface for DOI and PMID
