+++ api/src/main/java/org/vivoweb/webapp/createandlink/crossref/CrossrefNativeAPI.java
@@ -1,0 +1,365 @@
+package org.vivoweb.webapp.createandlink.crossref;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import edu.cornell.mannlib.vitro.webapp.utils.http.HttpClientFactory;
+import org.apache.commons.lang3.ArrayUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.apache.http.HttpResponse;
+import org.apache.http.client.HttpClient;
+import org.apache.http.client.methods.HttpGet;
+import org.vivoweb.webapp.createandlink.Citation;
+import org.vivoweb.webapp.createandlink.CreateAndLinkUtils;
+import org.vivoweb.webapp.createandlink.ResourceModel;
+import org.vivoweb.webapp.createandlink.utils.HttpReader;
+import org.vivoweb.webapp.createandlink.utils.StringArrayDeserializer;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
+/**
+ * Interface to CrossRef's native API
+ */
+@JsonIgnoreProperties
+public class CrossrefNativeAPI {
+    private static final Log log = LogFactory.getLog(CrossrefNativeAPI.class);
+
+    // API endpoint address
+    private static final String CROSSREF_API      = "http://api.crossref.org/works/";
+
+    /**
+     * Find the DOI in CrossRef, filling the citation object
+     *
+     * @param id
+     * @param citation
+     * @return
+     */
+    public String findInExternal(String id, Citation citation) {
+        // Get JSON from the CrossRef API
+        String json = readUrl(CROSSREF_API + id);
+
+        if (StringUtils.isEmpty(json)) {
+            return null;
+        }
+
+        CrossrefResponse response = null;
+        try {
+            ObjectMapper objectMapper = new ObjectMapper();
+            objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+            response = objectMapper.readValue(json, CrossrefResponse.class);
+        } catch (IOException e) {
+            log.error("Unable to read JSON value", e);
+        }
+        if (response == null || response.message == null) {
+            return null;
+        }
+
+        // The CrossRef API sometimes gives a false record when the DOI deosn't exist
+        // So ensure that the response we got contains the DOI we asked for
+        if (!id.equalsIgnoreCase(response.message.DOI)) {
+            return null;
+        }
+
+        // Map the fields from the CrossRef response to the Citation object
+
+        citation.DOI = id;
+        citation.type = normalizeType(response.message.type);
+
+        if (!ArrayUtils.isEmpty(response.message.title)) {
+            citation.title = response.message.title[0];
+        }
+
+        if (!ArrayUtils.isEmpty(response.message.containerTitle)) {
+            for (String journal : response.message.containerTitle) {
+                if (citation.journal == null || citation.journal.length() < journal.length()) {
+                    citation.journal = journal;
+                }
+            }
+        }
+
+        if (response.message.author != null) {
+            List<Citation.Name> authors = new ArrayList<>();
+            for (CrossrefResponse.ResponseModel.Author author : response.message.author) {
+                Citation.Name citationAuthor = new Citation.Name();
+                citationAuthor.name = CreateAndLinkUtils.formatAuthorString(author.family, author.given);
+                authors.add(citationAuthor);
[…]
