### 2018-12-05 91f75df
[VIVO-1415] Claiming interface for DOI and PMID
