+++ api/src/main/java/org/vivoweb/webapp/createandlink/ExternalIdentifiers.java
@@ -1,0 +1,7 @@
+package org.vivoweb.webapp.createandlink;
+
+public class ExternalIdentifiers {
+    public String DOI;
+    public String PubMedID;        // http://www.ncbi.nlm.nih.gov/pmc/utils/idconv/v1.0/?ids=23193287&format=json
+    public String PubMedCentralID; // http://www.ncbi.nlm.nih.gov/pmc/utils/idconv/v1.0/?ids=PMC3531190&format=json
+}
