### 2018-12-05 9441820
Allow http://doi.org urls in the claim process
