### 2018-12-06 cfe65e5
Fix PubMed import and licensing
