+++ api/src/main/java/org/vivoweb/webapp/createandlink/crossref/CrossrefCiteprocJSONModel.java
@@ -1,3 +1,5 @@
+/* $This file is distributed under the terms of the license in /doc/license.txt$ */
+
 package org.vivoweb.webapp.createandlink.crossref;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
