### 2018-12-07 35849f2
Check Editor already claimed, error for no citation, PUMCIDs fixed
