+++ api/src/main/java/org/vivoweb/webapp/controller/freemarker/CreateAndLinkResourceController.java
@@ -429,7 +429,6 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
                                 citation.externalProvider = externalProvider;
                             }
                         }
-
                     }
 
                     // Guess which author in the available metadata is the user claiming the work
@@ -441,6 +440,9 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
 
                         // Increment the count of processed identifiers
                         idCount++;
+                    } else {
+                        citation.showError = true;
+                        citations.add(citation);
                     }
                 }
             }
@@ -890,7 +892,21 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
 
         // Add a DOI
         if (!StringUtils.isEmpty(resourceModel.DOI)) {
-            work.addProperty(model.createProperty(BIBO_DOI), resourceModel.DOI.toLowerCase());
+            // Normalise the DOI according to link guidelines
+            String doi = resourceModel.DOI.toLowerCase();
+            if (doi.startsWith("http://dx.doi.org/")) {
+                doi = "https://doi.org/" + doi.substring(18);
+            } else if (doi.startsWith("https://dx.doi.org/")) {
+                doi = "https://doi.org/" + doi.substring(18);
+            } else if (doi.startsWith("http://doi.org/")) {
+                doi = "https://doi.org/" + doi.substring(15);
+            } else if (!doi.startsWith("http")) {
+                doi = "https://doi.org/" + doi;
+            }
+
+            if (doi.startsWith("https://doi.org/")) {
+                work.addProperty(model.createProperty(BIBO_DOI), doi);
+            }
         }
 
         // Add a PubMed ID
@@ -1459,21 +1475,21 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
                             }
                             break;
 
-                        // Relationships - we are really interested in authors
+                        // Relationships - we are really interested in authors and editors
                         case VIVO_RELATEDBY:
                             // Get the relationship context
                             Resource relationship = stmt.getResource();
                             if (relationship != null) {
                                 Integer rank = null;
 
-                                // If it isn't an authorship, skip it
-                                if (!isResourceOfType(relationship, VIVO_AUTHORSHIP)) {
+                                // If it isn't an authorship or editorship, skip it
+                                if (!isResourceOfType(relationship, VIVO_AUTHORSHIP) &&
+                                    !isResourceOfType(relationship, VIVO_EDITORSHIP)) {
                                     break;
                                 }
 
-                                // Now loop over the properties of the authorship context
-                                Citation.Name newAuthor = null;
-                                Resource authorResource = null;
+                                // Now loop over the properties of the author/editorship context
+                                Resource personResource = null;
                                 StmtIterator relationshipIter = relationship.listProperties();
                                 try {
                                     while (relationshipIter.hasNext()) {
@@ -1483,7 +1499,7 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
                                             case VIVO_RELATES:
                                                 // If it isn't pointing to the resource, it must be pointing to a person
                                                 if (!vivoUri.equals(relationshipStmt.getResource().getURI())) {
-                                                    authorResource = relationshipStmt.getResource();
+                                                    personResource = relationshipStmt.getResource();
                                                 }
                                                 break;
 
@@ -1497,80 +1513,86 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
                                     relationshipIter.close();
                                 }
 
+                                Citation.Name newAuthor = null;
+
                                 // If we've got an author
-                                if (authorResource != null) {
+                                if (personResource != null) {
                                     // If the author is the user, then they have already claimed this publication
[…]
