### 2018-12-07 f9ef9e3
Check Editor already claimed, error for no citation, PUMCIDs fixed
