+++ api/src/main/java/org/vivoweb/webapp/controller/freemarker/CreateAndLinkResourceController.java
@@ -4,6 +4,7 @@ package org.vivoweb.webapp.controller.freemarker;
 
 import edu.cornell.mannlib.vedit.beans.LoginStatusBean;
 import edu.cornell.mannlib.vitro.webapp.auth.permissions.SimplePermission;
+import edu.cornell.mannlib.vitro.webapp.auth.policy.PolicyHelper;
 import edu.cornell.mannlib.vitro.webapp.auth.requestedAction.AuthorizationRequest;
 import edu.cornell.mannlib.vitro.webapp.beans.Individual;
 import edu.cornell.mannlib.vitro.webapp.beans.SelfEditingConfiguration;
@@ -251,19 +252,27 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
             person = individualDao.getIndividualByURI(profileUri);
         }
 
-        // If we haven't got a valid person
-        if (person == null) {
-            // Get the currently logged in user
-            UserAccount loggedInAccount = LoginStatusBean.getCurrentUser(vreq);
-            SelfEditingConfiguration sec = SelfEditingConfiguration.getBean(vreq);
+        boolean isProfileUriForLoggedIn = false;
 
-            // Find the profile(s) associated with this user
-            List<Individual> assocInds = sec.getAssociatedIndividuals(vreq.getWebappDaoFactory().getIndividualDao(), loggedInAccount.getExternalAuthId());
-            if (!assocInds.isEmpty()) {
+        // Get the currently logged in user
+        UserAccount loggedInAccount = LoginStatusBean.getCurrentUser(vreq);
+        SelfEditingConfiguration sec = SelfEditingConfiguration.getBean(vreq);
+
+        // Find the profile(s) associated with this user
+        List<Individual> assocInds = sec.getAssociatedIndividuals(vreq.getWebappDaoFactory().getIndividualDao(), loggedInAccount.getExternalAuthId());
+        if (!assocInds.isEmpty()) {
+            if (person == null) {
                 // If we have associated profiles, ensure that a valid person profile really does exist
                 profileUri = assocInds.get(0).getURI();
                 if (!StringUtils.isEmpty(profileUri)) {
                     person = individualDao.getIndividualByURI(profileUri);
+                    isProfileUriForLoggedIn = true;
+                }
+            } else if (!StringUtils.isEmpty(profileUri)){
+                for (Individual ind : assocInds) {
+                    if (ind.getURI().equalsIgnoreCase(profileUri)) {
+                        isProfileUriForLoggedIn = true;
+                    }
                 }
             }
         }
@@ -273,7 +282,12 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
             return new TemplateResponseValues("unknownProfile.ftl");
         }
 
-        // Create a map of common values to pass to the templates
+        // If the profile isn't associated with the logged in user, check that we have back end editing privileges
+        if (!isProfileUriForLoggedIn && !PolicyHelper.isAuthorizedForActions(vreq, SimplePermission.DO_BACK_END_EDITING.ACTION)) {
+            return new TemplateResponseValues("unauthorizedForProfile.ftl");
+        }
+
+            // Create a map of common values to pass to the templates
         Map<String, Object> templateValues = new HashMap<>();
         templateValues.put("link", profileUri);
         templateValues.put("label", provider.getLabel());
@@ -374,8 +388,8 @@ public class CreateAndLinkResourceController extends FreemarkerHttpServlet {
             Set<String> remainderIds = new HashSet<>();
             List<Citation> citations = new ArrayList<>();
 
-            // Split the passed IDs into a parseable array (separated by whitespace, comma or semicolon)
[…]
