### 2018-12-10 787ddb5
[VIVO-1641] REPLACE jena afn with cross-platform regex
