+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/controller/freemarker/ManagePublicationsForIndividualController.java
@@ -69,7 +69,6 @@ public class ManagePublicationsForIndividualController extends FreemarkerHttpSer
         + "PREFIX core: <http://vivoweb.org/ontology/core#> \n"
         + "PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#> \n"
         + "PREFIX vitro: <http://vitro.mannlib.cornell.edu/ns/vitro/0.7#> \n"
-        + "PREFIX afn: <http://jena.apache.org/ARQ/function#> \n"
         + "SELECT DISTINCT ?subclass ?authorship (str(?label) as ?title) ?pub ?hideThis WHERE { \n"
         + "    ?subject core:relatedBy ?authorship . \n"
         + "    ?authorship a core:Authorship  . \n"
