### 2018-12-16 df48788
Update DOI resolver URI for content negotiation
