### 2018-12-16 93f7f55
Add VCard indexer extension for VCards in Authorships/Editorships
