+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/searchindex/extensions/VCardsAcrossContextNodes.java
@@ -1,0 +1,471 @@
+/* $This file is distributed under the terms of the license in LICENSE$ */
+
+package edu.cornell.mannlib.vitro.webapp.searchindex.extensions;
+
+import edu.cornell.mannlib.vitro.webapp.beans.Individual;
+import edu.cornell.mannlib.vitro.webapp.beans.VClass;
+import edu.cornell.mannlib.vitro.webapp.modelaccess.ContextModelAccess;
+import edu.cornell.mannlib.vitro.webapp.modules.searchEngine.SearchInputDocument;
+import edu.cornell.mannlib.vitro.webapp.rdfservice.RDFService;
+import edu.cornell.mannlib.vitro.webapp.rdfservice.RDFServiceException;
+import edu.cornell.mannlib.vitro.webapp.searchindex.documentBuilding.DocumentModifier;
+import edu.cornell.mannlib.vitro.webapp.searchindex.indexing.IndexingUriFinder;
+import edu.cornell.mannlib.vitro.webapp.utils.configuration.ContextModelsUser;
+import edu.cornell.mannlib.vitro.webapp.utils.configuration.Property;
+import edu.cornell.mannlib.vitro.webapp.utils.configuration.Validation;
+import edu.cornell.mannlib.vitro.webapp.visualization.constants.QueryConstants;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.apache.jena.rdf.model.Model;
+import org.apache.jena.rdf.model.ModelFactory;
+import org.apache.jena.rdf.model.ResourceFactory;
+import org.apache.jena.rdf.model.Statement;
+import org.apache.jena.rdf.model.StmtIterator;
+import org.apache.jena.vocabulary.RDFS;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import static edu.cornell.mannlib.vitro.webapp.search.VitroSearchTermNames.ALLTEXT;
+import static edu.cornell.mannlib.vitro.webapp.search.VitroSearchTermNames.ALLTEXTUNSTEMMED;
+import static edu.cornell.mannlib.vitro.webapp.utils.sparqlrunner.SparqlQueryRunner.createSelectQueryContext;
+
+/**
+ * If an individual has context nodes then the search document for that
+ * individual should include the labels of the partners across those nodes. The
+ * labels will be added to the ALLTEXT and ALLTEXTUNSTEMMED fields.
+ * 
+ * We must specify what property leads to a context node (incoming), and what
+ * property leads from a context node (outgoing). We may add restrictions to say
+ * that this only applies to individuals of certain types. We may also restrict
+ * the type of the applicable context nodes.
+ * 
+ * An instance of this class acts as both a DocumentModifier and an
+ * IndexingUriFinder:
+ * 
+ * As a DocumentModifier, it looks across approved context nodes to fetch the
+ * labels of the partners.
+ * 
+ * As an IndexingUriFinder, it recognizes that this relationship can be changed
+ * by a change to a "label" statement, or to a "relates" property, and finds all
+ * partners as candidates for reindexing.
+ * 
+ * <pre>
+ * Configuration:
+ *     rdfs:label -- Optional. Appears in the timings and debug statements.
+ *     :hasIncomingProperty -- Required. Property leading to the context node.
[…]
