### 2018-12-17 f970459
ORCID Internationalization, allow Step 2 to be disabled if only a public API key
