### 2018-12-19 031e0c5
[VIVO-1415] Add selectable list of publication types on import
