### 2018-12-19 85fc42a
[VIVO-1415] Add selectable list of publication types on import
