### 2018-12-19 aa0a769
[VIVO-1415] Try to coalesce publisher by it's name
