### 2019-01-16 #1182 @whedon
[REVIEW]: VIVO: a system for research discovery
**Submitting author:** @mconlon17 (<a href="http://orcid.org/0000-0002-1304-8447">Michael Conlon</a>)
**Repository:** <a href="https://github.com/vivo-project/VIVO" target ="_blank">https://github.com/vivo-project/VIVO</a>
**Version:** v1.10.0
