### 2019-01-16 7150d06
Fix NullPointer in visualisations
