### 2019-02-08 222ddb6
Fix publication counts in histogram
