+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/visualization/personpubcount/CumulativeCountRequestHandler.java
@@ -3,6 +3,7 @@
 package edu.cornell.mannlib.vitro.webapp.visualization.personpubcount;
 
 import edu.cornell.mannlib.vitro.webapp.auth.requestedAction.AuthorizationRequest;
+import edu.cornell.mannlib.vitro.webapp.config.ConfigurationProperties;
 import edu.cornell.mannlib.vitro.webapp.controller.VitroRequest;
 import edu.cornell.mannlib.vitro.webapp.controller.freemarker.responsevalues.ResponseValues;
 import edu.cornell.mannlib.vitro.webapp.visualization.exceptions.MalformedQueryParametersException;
@@ -11,17 +12,49 @@ import edu.cornell.mannlib.vitro.webapp.visualization.visutils.QueryRunner;
 import edu.cornell.mannlib.vitro.webapp.visualization.visutils.VisualizationRequestHandler;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.apache.jena.query.Dataset;
 
+import java.util.AbstractMap.SimpleEntry;
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class CumulativeCountRequestHandler implements VisualizationRequestHandler {
+    private static final Log log = LogFactory.getLog(CumulativeCountRequestHandler.class);
+
+    // Flag to say whether we have attempted to load column definitions from the runtime.properties
+    private static boolean loadedColumnDefinitions = false;
+
+    // Default configuration for pub lication groups and the types they contain
+    private static String[] pubsGroups = { "Articles", "Books" };
+
+    // Default types for each publications group
+    private static Map<String, String[]> groupTypeMap = Stream.of(
+        new SimpleEntry<String, String[]>(
+                    "Articles", new String[] {
+                            "http://purl.org/ontology/bibo/AcademicArticle",
+                            "http://purl.org/ontology/bibo/Article"
+                        }
+                ),
+        new SimpleEntry<String, String[]>(
+                    "Books", new String[] {
+                            "http://purl.org/ontology/bibo/Book",
+                            "http://purl.org/ontology/bibo/BookSection",
+                            "http://purl.org/ontology/bibo/Chapter",
+                            "http://purl.org/ontology/bibo/EditedBook"
+                        }
+                )
+    ).collect(Collectors.toMap(SimpleEntry::getKey, SimpleEntry::getValue));
+
     @Override
     public AuthorizationRequest getRequiredPrivileges() {
         return null;
@@ -39,16 +72,19 @@ public class CumulativeCountRequestHandler implements VisualizationRequestHandle
 
     @Override
     public Object generateAjaxVisualization(VitroRequest vitroRequest, Log log, Dataset dataSource) throws MalformedQueryParametersException {
+        // Ensure that we have loaded any configured definitions
+        loadColumnDefintions(vitroRequest);
+
         String personURI = vitroRequest.getParameter("uri");
 
-        QueryRunner<Set<Activity>> queryManager = new PersonPublicationCountQueryRunner(
+        QueryRunner<Collection<Activity>> queryManager = new PersonPublicationCountQueryRunner(
                 personURI,
                 vitroRequest.getRDFService(),
                 log);
 
-        Set<Activity> authorDocuments = queryManager.getQueryResult();
+        Collection<Activity> authorDocuments = queryManager.getQueryResult();
 
[…]
