### 2019-03-13 #1182 @labarba
👋 @sonal-raj — I see from a [comment](https://github.com/openjournals/joss-reviews/issues/943#issuecomment-454131124) in the Pre-review issue that you agreed to get to this review by the end of January. We have a few items ticked off on your review checklist, so it looks like you did get started! Do you have an idea when you might get finished, though?
