### 2019-03-14 #1182 @sonal-raj
The VIVO software has been reviewed against all the recommended criteria in the checklist. The installation, functionality and documentation are sound and the current version is operational without issues. I recommend the acceptance of this submission.
