### 2019-03-19 29ccf74
Update link to installation doc in README.md
