### 2019-03-20 8511073
[VIVO-1641] Replace afn:localname / afn:namespace with cross-platform equivalent (#96)
* Update listViewConfig-default.xml
* 1641 repalce localname
* [VIVO-1641] REPLACE jena afn with cross-platform regex

Resolves https://jira.duraspace.org/browse/VIVO-1641
