+++ README.md
@@ -1,6 +1,6 @@
 # VIVO: Connect, Share, Discover
 
-[![Build Status](https://travis-ci.org/vivo-project/VIVO.png?branch=develop)](https://travis-ci.org/vivo-project/VIVO)
+[![Build Status](https://travis-ci.org/vivo-project/VIVO.png?branch=develop)](https://travis-ci.org/vivo-project/VIVO) [![DOI][![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2639714.svg)](https://doi.org/10.5281/zenodo.2639713)
 
 VIVO is an open source semantic web tool for research discovery -- finding people and the research they do.
 
