### 2019-04-15 #1182 @arfon
@mconlon17 - there are a couple of issues with your paper. One of the grant numbers looks to be a placeholder (`xxxxxxx`) - could you please address that.

Also, none of the references are included in the body text which means they won't be compiled into the paper pdf. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
