### 2019-04-25 bd6e532
Remove trailing whitespace throughout project
