+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/controller/AboutQrCodesController.java
@@ -1,6 +1,6 @@
 /* $This file is distributed under the terms of the license in LICENSE$ */
 
-package edu.cornell.mannlib.vitro.webapp.controller; 
+package edu.cornell.mannlib.vitro.webapp.controller;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -18,7 +18,7 @@ import javax.servlet.annotation.WebServlet;
 
 @WebServlet(name = "AboutQrCodesController", urlPatterns = {"/qrcode/about"})
 public class AboutQrCodesController extends FreemarkerHttpServlet {
-    
+
     private static final long serialVersionUID = 1L;
     private static final Log log = LogFactory.getLog(ExportQrCodeController.class);
     private static final String TEMPLATE_DEFAULT = "aboutQrCodes.ftl";
@@ -27,7 +27,7 @@ public class AboutQrCodesController extends FreemarkerHttpServlet {
     protected ResponseValues processRequest(VitroRequest vreq) {
         try {
             Map<String, Object> body = new HashMap<String, Object>();
-            
+
             return new TemplateResponseValues(TEMPLATE_DEFAULT, body);
         } catch (Throwable e) {
             log.error(e, e);
