+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/edit/n3editing/configuration/generators/ManageLabelsForPersonGenerator.java
@@ -46,7 +46,7 @@ import edu.cornell.mannlib.vitro.webapp.web.templatemodels.individual.DataProper
 /**
  * This generator is specifically for handling labels for a FOAF Person individual and is an object property form.
  *This allows the page to show all the labels for a particular individual and sets up code
- *enabling the addition of a new label.  Links on the page will allow for removal or editing of a given label. 
+ *enabling the addition of a new label.  Links on the page will allow for removal or editing of a given label.
  */
 public class ManageLabelsForPersonGenerator extends BaseEditConfigurationGenerator implements EditConfigurationGenerator {
     public static Log log = LogFactory.getLog(ManageLabelsForIndividualGenerator.class);
@@ -68,13 +68,13 @@ public class ManageLabelsForPersonGenerator extends BaseEditConfigurationGenerat
         //be employed when the 'add' button is used, we will set this is an object property form
         //although label will mean we need to add a data property statement as well
         //URL to return to is the same page once addition is complete
-        initObjectPropForm(config, vreq);  
+        initObjectPropForm(config, vreq);
 
-        
+
         this.setUrlToReturnTo(config, vreq);
 
         config.setSubjectUri(EditConfigurationUtils.getSubjectUri(vreq));
-        
+
         setVarNames(config);
         //config.setDatapropKey( EditConfigurationUtils.getDataHash(vreq) );
         //Add n3, fields, etc. in the case where the user wants to add a label
@@ -86,10 +86,10 @@ public class ManageLabelsForPersonGenerator extends BaseEditConfigurationGenerat
     	this.setUrisAndLiteralsInScope(config);
         this.setFields(config, vreq, EditConfigurationUtils
 				.getPredicateUri(vreq));
-        
+
         //Get existing labels
         //this.initExistingLabels(config, vreq);
-        
+
         //Add form specific data used to populate template
         addFormSpecificData(config, vreq);
         //This preprocessor handles getting the correct label language and putting the attribute on the label
@@ -97,24 +97,24 @@ public class ManageLabelsForPersonGenerator extends BaseEditConfigurationGenerat
 			   new ManageLabelsForPersonPreprocessor(config));
  	   //This will handle generating the label from the first name, middle, and last names and also make sure to associate
  	   //a language with that label
-       config.addModelChangePreprocessor(new FoafNameToRdfsLabelPreprocessor());        
+       config.addModelChangePreprocessor(new FoafNameToRdfsLabelPreprocessor());
 
         prepare(vreq, config);
         return config;
     }
-    
+
     /**With ISF Changes**/
     //For addition of a label, with ISF changes, the name is now linked to a vcard which  in turn is linked to a "fullname" that then has first/middle/last names
-    
+
     private void addNewResources(EditConfigurationVTwo config) {
     	config.addNewResource("fullName", DEFAULT_NS_FOR_NEW_RESOURCE);
         config.addNewResource("individualVcard", DEFAULT_NS_FOR_NEW_RESOURCE);
     }
-  
+
     private void setUrlToReturnTo(EditConfigurationVTwo editConfiguration, VitroRequest vreq) {
-		editConfiguration.setUrlPatternToReturnTo(EditConfigurationUtils.getFormUrlWithoutContext(vreq));		
+		editConfiguration.setUrlPatternToReturnTo(EditConfigurationUtils.getFormUrlWithoutContext(vreq));
 	}
-    
+
[…]
