+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/edit/n3editing/configuration/generators/AddOrcidIdToPersonGenerator.java
@@ -17,46 +17,46 @@ import edu.cornell.mannlib.vitro.webapp.edit.n3editing.configuration.validators.
 public class AddOrcidIdToPersonGenerator extends VivoBaseGenerator implements
         EditConfigurationGenerator {
     private Log log = LogFactory.getLog(AddOrcidIdToPersonGenerator.class);
-       
+
     @Override
     public EditConfigurationVTwo getEditConfiguration(VitroRequest vreq,
             HttpSession session) throws Exception {
-        
+
         EditConfigurationVTwo conf = new EditConfigurationVTwo();
-        
+
         initBasics(conf, vreq);
         initPropertyParameters(vreq, session, conf);
-        initObjectPropForm(conf, vreq);  
-        
+        initObjectPropForm(conf, vreq);
+
         conf.setTemplate("addOrcidIdToPerson.ftl");
-        
+
         conf.setVarNameForSubject("person");
         conf.setVarNameForPredicate("predicate");
         conf.setVarNameForObject("orcidId");
-        
+
         conf.setN3Required( Arrays.asList( n3ForOrcidId ) );
-                        
+
         conf.setUrisOnform(Arrays.asList("orcidId"));
-        
+
         conf.addSparqlForExistingUris("orcidId", orcidIdQuery);
 
-        conf.addField( new FieldVTwo().                        
+        conf.addField( new FieldVTwo().
                 setName("orcidId").
                 setValidators( list("nonempty") ));
-        
+
         conf.addValidator(new AntiXssValidation());
-        
+
         prepare(vreq, conf);
         return conf;
     }
 
     /* N3 assertions  */
 
-    final static String n3ForOrcidId = 
+    final static String n3ForOrcidId =
         "@prefix owl: <http://www.w3.org/2002/07/owl#> .\n"+
         "?person <http://vivoweb.org/ontology/core#orcidId> ?orcidId . \n" +
-        "?orcidId a owl:Thing . " ;    
-    
+        "?orcidId a owl:Thing . " ;
+
     /* Queries for editing an existing entry */
 
     final static String orcidIdQuery =

[…]
