### 2019-04-26 f7b67ce
Remove Maven file
