+++ check.txt
@@ -1,810 +1,0 @@
-[INFO] Scanning for projects...
-[INFO] ------------------------------------------------------------------------
-[INFO] Reactor Build Order:
-[INFO]
-[INFO] Vitro
-[INFO] Vitro Dependencies
-[INFO] Vitro API
-[INFO] VIVO
-[INFO] VIVO API
-[INFO] Vitro Web App
-[INFO] VIVO Web App
-[INFO] Vitro Home
-[INFO] VIVO Home
-[INFO] Vitro Solr App
-[INFO] Vitro Installer
-[INFO] Vitro Install Home
-[INFO] Vitro Install Solr App
-[INFO] Vitro Install Web App
-[INFO] VIVO Installer
-[INFO] VIVO Install Home
-[INFO] VIVO Install Solr App
-[INFO] VIVO Install Web App
-[INFO]
-[INFO] ------------------------------------------------------------------------
-[INFO] Building Vitro 1.11.0-SNAPSHOT
-[INFO] ------------------------------------------------------------------------
-[INFO]
-[INFO] --- maven-clean-plugin:3.0.0:clean (default-clean) @ vitro-project ---
-[INFO]
-[INFO] --- buildnumber-maven-plugin:1.4:create (default) @ vitro-project ---
-[INFO] ShortRevision tag detected. The value is '7'.
-[INFO] Executing: /bin/sh -c cd '/Users/benjamingross/Documents/git/Vitro' && 'git' 'rev-parse' '--verify' '--short=7' 'HEAD'
-[INFO] Working directory: /Users/benjamingross/Documents/git/Vitro
-[INFO] Storing buildNumber: 129e121 at timestamp: 1544232832342
-[INFO] Storing buildScmBranch: bugfix/VIVO-1610
-[INFO]
-[INFO] --- maven-install-plugin:2.4:install (default-install) @ vitro-project ---
-[INFO] Installing /Users/benjamingross/Documents/git/Vitro/pom.xml to /Users/benjamingross/.m2/repository/org/vivoweb/vitro-project/1.11.0-SNAPSHOT/vitro-project-1.11.0-SNAPSHOT.pom
-[INFO]
-[INFO] ------------------------------------------------------------------------
-[INFO] Building Vitro Dependencies 1.11.0-SNAPSHOT
-[INFO] ------------------------------------------------------------------------
-[INFO]
-[INFO] --- maven-clean-plugin:3.0.0:clean (default-clean) @ vitro-dependencies ---
-[INFO]
-[INFO] --- buildnumber-maven-plugin:1.4:create (default) @ vitro-dependencies ---
-[INFO] ShortRevision tag detected. The value is '7'.
-[INFO] Executing: /bin/sh -c cd '/Users/benjamingross/Documents/git/Vitro/dependencies' && 'git' 'rev-parse' '--verify' '--short=7' 'HEAD'
-[INFO] Working directory: /Users/benjamingross/Documents/git/Vitro/dependencies
-[INFO] Storing buildNumber: 129e121 at timestamp: 1544232832498
-[INFO] Storing buildScmBranch: bugfix/VIVO-1610
-[INFO]
-[INFO] --- maven-install-plugin:2.4:install (default-install) @ vitro-dependencies ---
[…]
