### 2019-04-30 86e5da7
Add i18n label support, minor code review tweaks
