### 2019-05-03 87092ce
[VIVO-1690] - Remove trailing whitespace throughout project (#126)
* Remove trailing whitespace throughout project

Resolves: https://jira.duraspace.org/browse/VIVO-1690
** Along with: https://github.com/vivo-project/Vitro/pull/117
