### 2019-05-20 #1182 @arfon
I should have updated here to say I emailed the author last night about this and he replied immediately. I would expect an update shortly.
