### 2019-06-05 e96cf65
Improve Self Editing Performance
