+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/edit/n3editing/configuration/generators/ManageLabelsForPersonGenerator.java
@@ -317,11 +317,11 @@ public class ManageLabelsForPersonGenerator extends BaseEditConfigurationGenerat
 	private Object isEditable(VitroRequest vreq, EditConfigurationVTwo config) {
 		Individual individual = EditConfigurationUtils.getIndividual(vreq, config.getSubjectUri());
 		AddDataPropertyStatement adps = new AddDataPropertyStatement(
-				vreq.getJenaOntModel(), individual.getURI(),
+				vreq, vreq.getJenaOntModel(), individual.getURI(),
 				SOME_URI, SOME_LITERAL);
 
 		AddObjectPropertyStatement aops = new AddObjectPropertyStatement(
-				vreq.getJenaOntModel(), individual.getURI(),
+				vreq, vreq.getJenaOntModel(), individual.getURI(),
 				SOME_PREDICATE, SOME_URI);
     	return PolicyHelper.isAuthorizedForActions(vreq, adps.or(aops));
 	}
