### 2019-06-05 7f4651c
Improve Self Editing Performance
