### 2019-07-01 #1182 @kyleniemeyer
Hi @mconlon17, just checking in about your updates to the paper.
