### 2019-07-02 bbfda0e
Activity.java whitespace fixes to resolve conflict
