### 2019-07-02 beb8ba0
Clean up whitespace on PersonPublicationCountRequestHandler.java
