+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/visualization/personpubcount/PersonPublicationCountRequestHandler.java
@@ -28,7 +28,7 @@ import edu.cornell.mannlib.vitro.webapp.visualization.visutils.UtilityFunctions;
 import edu.cornell.mannlib.vitro.webapp.visualization.visutils.VisualizationRequestHandler;
 
 /**
- * 
+ *
  * This request handler is used to serve the content related to an individual's
  * publications over the years like, 1. Sprakline representing this 2. An entire
  * page dedicated to the sparkline vis which will also have links to download
@@ -37,7 +37,7 @@ import edu.cornell.mannlib.vitro.webapp.visualization.visutils.VisualizationRequ
  * years. 4. Downloadable PDf file containing the publications content, among
  * other things. Currently this is disabled because the feature is half-baked.
  * We plan to activate this in the next major release.
- * 
+ *
  * @author cdtank
  */
 public class PersonPublicationCountRequestHandler implements
@@ -64,7 +64,7 @@ VisualizationRequestHandler {
         /* personPublicationCountDynamicActivator.ftl template needs to know which is the requesting template. */
         String requestingTemplate = vitroRequest
             					.getParameter(
-										VisualizationFrameworkConstants.REQUESTING_TEMPLATE_KEY);  
+										VisualizationFrameworkConstants.REQUESTING_TEMPLATE_KEY);
 
 		QueryRunner<Collection<Activity>> queryManager = new PersonPublicationCountQueryRunner(
 															personURI,
@@ -77,7 +77,7 @@ VisualizationRequestHandler {
 		 * Create a map from the year to number of publications. Use the
 		 * BiboDocument's parsedPublicationYear to populate the data.
 		 */
-		Map<String, Integer> yearToPublicationCount = 
+		Map<String, Integer> yearToPublicationCount =
 				UtilityFunctions.getYearToActivityCount(authorDocuments);
 
 		boolean shouldVIVOrenderVis = false;
@@ -96,11 +96,11 @@ VisualizationRequestHandler {
 		 * Computations required to generate HTML for the sparkline & related
 		 * context.
 		 */
-		PersonPublicationCountVisCodeGenerator visualizationCodeGenerator = 
+		PersonPublicationCountVisCodeGenerator visualizationCodeGenerator =
 				new PersonPublicationCountVisCodeGenerator(
-						personURI, 
-						visMode, 
-						visContainer, 
+						personURI,
+						visMode,
+						visContainer,
 						yearToPublicationCount,
 						log);
 
@@ -110,12 +110,12 @@ VisualizationRequestHandler {
 				shouldVIVOrenderVis, requestingTemplate);
 
 	}
-	
+
 	@Override
 	public ResponseValues generateVisualizationForShortURLRequests(
 			Map<String, String> parameters, VitroRequest vitroRequest, Log log,
 			Dataset dataSource) throws MalformedQueryParametersException {
-		throw new UnsupportedOperationException("Person Publication Count Visualization does not provide " 
+		throw new UnsupportedOperationException("Person Publication Count Visualization does not provide "
 					+ "Short URL Response.");
 	}
 
@@ -137,7 +137,7 @@ VisualizationRequestHandler {
 		 * Create a map from the year to number of publications. Use the
 		 * BiboDocument's parsedPublicationYear to populate the data.
 		 */
-		Map<String, Integer> yearToPublicationCount = 
+		Map<String, Integer> yearToPublicationCount =
 				UtilityFunctions.getYearToActivityCount(authorDocuments);
 
 		String authorName = ((PersonPublicationCountQueryRunner) queryManager).getAuthorName();
[…]
