### 2019-07-15 801304f
Update paper.bib
Added dates to websites according the respective website, checked links, updated urldates, added "author = {VIVO Project}" to the references of "About Vitro" and "VIVO-ISF-Ontology", changes "Vivo-isf-ontology" to "VIVO-ISF-Ontology".
