+++ paper/paper.bib
@@ -14,20 +14,23 @@
 }
 
 @misc{vivo-isf-ontology:_2018,
-	title = {vivo-isf-ontology: {The} "{VIVO}-{ISF} {Ontology}" is an {OWL}2 representation of the {VIVO}-{ISF} {Data} {Standard}},
-	shorttitle = {vivo-isf-ontology},
+        author = {VIVO Project},
+	title = {VIVO-ISF-ontology: {The} "{VIVO}-{ISF} {Ontology}" is an {OWL}2 representation of the {VIVO}-{ISF} {Data} {Standard}},
+	shorttitle = {VIVO-ISF-Ontology},
 	url = {https://github.com/openrif/vivo-isf-ontology},
-	urldate = {2018-07-14},
+	urldate = {2019-07-15},
 	publisher = {OpenRIF},
 	month = apr,
-	year = {2018},
+	year = 2018,
 	note = {original-date: 2014-09-08T21:40:43Z}
 }
 
 @misc{about_vitro:2018,
+        author = {VIVO Project},
 	title = {About {Vitro} {\textbar} {VIVO}},
 	url = {http://vivoweb.org/info/about-vitro},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2019,
 }
 
 @article{verborgh_triple_2016,
@@ -50,49 +53,57 @@
 @misc{jfact_2018,
 	title = {{JFact} {DL} {Reasoner}},
 	url = {http://jfact.sourceforge.net/},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2018
 }
 
 @misc{solr_2018,
 	title = {Apache {Solr} -},
 	url = {http://lucene.apache.org/solr/},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2019
 }
 
 @misc{sparql_2018,
 	title = {{SPARQL} {Query} {Language} for {RDF}},
 	url = {https://www.w3.org/TR/rdf-sparql-query/},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2008
 }
 
 @misc{jena_2018,
 	title = {Apache {Jena} -},
 	url = {https://jena.apache.org/},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2011
 }
 
 @misc{jena_tdb_2018,
 	title = {Apache {Jena} - {Apache} {Jena} - {TDB}},
 	url = {https://jena.apache.org/documentation/tdb/},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2019
 }
 
 @misc{jena_sdb_2018,
 	title = {Apache {Jena} - {SDB} - persistent triple stores using relational databases},
 	url = {https://jena.apache.org/documentation/sdb/index.html},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2019
 }
 
 @misc{mysql_2018,
 	title = {{MySQL}},
 	url = {https://www.mysql.com/},
-	urldate = {2018-07-20}
+	urldate = {2019-07-15}
+	year = 2019
 }
 
 @misc{bostock_d3_2018,
 	title = {D3.js - {Data}-{Driven} {Documents}},
 	url = {https://d3js.org/},
 	abstract = {D3 is a JavaScript library for visualizing data with HTML, SVG, and CSS.},
-	urldate = {2018-07-20},
-	author = {Bostock, Mike}
+	urldate = {2019-07-15},
+	author = {Bostock, Mike},
+	year = 2015
 }
