### 2019-07-15 b7edca3
Update paper.bib
deleted comma
