+++ paper/paper.bib
@@ -30,7 +30,7 @@
 	title = {About {Vitro} {\textbar} {VIVO}},
 	url = {http://vivoweb.org/info/about-vitro},
 	urldate = {2019-07-15}
-	year = 2019,
+	year = 2019
 }
 
 @article{verborgh_triple_2016,
