### 2019-07-15 #1182 @arfon
@mconlon17 - this PR should allow your paper to compile: https://github.com/vivo-project/VIVO/pull/134
