### 2019-07-24 89bdf16
Update paper.bib (#134)
