+++ paper/paper.bib
@@ -29,7 +29,7 @@
         author = {VIVO Project},
 	title = {About {Vitro} {\textbar} {VIVO}},
 	url = {http://vivoweb.org/info/about-vitro},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2019
 }
 
@@ -53,49 +53,49 @@
 @misc{jfact_2018,
 	title = {{JFact} {DL} {Reasoner}},
 	url = {http://jfact.sourceforge.net/},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2018
 }
 
 @misc{solr_2018,
 	title = {Apache {Solr} -},
 	url = {http://lucene.apache.org/solr/},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2019
 }
 
 @misc{sparql_2018,
 	title = {{SPARQL} {Query} {Language} for {RDF}},
 	url = {https://www.w3.org/TR/rdf-sparql-query/},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2008
 }
 
 @misc{jena_2018,
 	title = {Apache {Jena} -},
 	url = {https://jena.apache.org/},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2011
 }
 
 @misc{jena_tdb_2018,
 	title = {Apache {Jena} - {Apache} {Jena} - {TDB}},
 	url = {https://jena.apache.org/documentation/tdb/},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2019
 }
 
 @misc{jena_sdb_2018,
 	title = {Apache {Jena} - {SDB} - persistent triple stores using relational databases},
 	url = {https://jena.apache.org/documentation/sdb/index.html},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2019
 }
 
 @misc{mysql_2018,
 	title = {{MySQL}},
 	url = {https://www.mysql.com/},
-	urldate = {2019-07-15}
+	urldate = {2019-07-15},
 	year = 2019
 }
 
