### 2019-07-25 #1182 @mconlon17
I have reviewed the proof.  All good.  Thanks to @hauschke for addressing comments.
